@@ -1,12 +1,13 @@
 {
   "name": "@loremipsum/mocking-hans",
   "author": "Kevin Hirczy",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "You don't mock the Hans, he's mocking you.",
   "scripts": {
     "example": "ts-node -r tsconfig-paths/register example/index.ts",
-    "test": "npm run lint",
-    "build": "npm run lint && tsc",
+    "test": "npm run lint && jest",
+    "test-coverage": "jest --coverage",
+    "build": "npm run test && tsc",
     "lint": "tslint 'src/**/*.ts'"
   },
   "main": "dist/index.js",
@@ -15,27 +16,44 @@
     "type": "git",
     "url": "git@projekte.loremipsum.at:li/dev/mocking-hans.git"
   },
-  "files": ["dist/**/*"],
+  "files": [
+    "dist/**/*"
+  ],
+  "bin": {
+    "hans": "./dist/console.js"
+  },
   "private": true,
   "dependencies": {
     "@types/express": "^4.16.1",
     "@types/faker": "^4.1.5",
-    "@types/node": "^10.12.10",
+    "@types/node": "^10.14.4",
+    "@types/socket.io": "^2.1.2",
     "@types/ws": "^6.0.1",
     "body-parser": "^1.18.3",
-    "chalk": "^2.4.1",
+    "chalk": "^2.4.2",
     "express": "^4.16.4",
+    "express-graphql": "^0.8.0",
     "faker": "^4.1.0",
+    "graphql": "^14.2.1",
     "jsontoxml": "^1.0.1",
     "morgan": "^1.9.1",
-    "reflect-metadata": "^0.1.12",
-    "socket.io": "^2.1.1",
-    "socket.io-client": "^2.1.1",
+    "reflect-metadata": "^0.1.13",
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.2.0",
     "ts-node": "^7.0.1",
-    "typescript": "^3.1.3",
-    "ws": "^6.1.2"
+    "ts-node-dev": "^1.0.0-pre.32",
+    "typescript": "^3.4.3",
+    "ws": "^6.2.1"
   },
   "devDependencies": {
+    "@types/jest": "^24.0.11",
+    "jest": "^24.7.1",
+    "ts-jest": "^24.0.2",
+    "tsconfig-paths": "^3.8.0",
     "tslint": "^5.15.0"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node"
   }
 }
