@@ -1,4 +1,5 @@
 export * from './HttpMethod';
-export * from './Metadata';
+export * from './MetadataKey';
 export * from './RouteDefinition';
 export * from './SocketDefinition';
+export * from './GraphqlRouteDefinition';
